+++ src/openassay/__init__.py
@@ -7,6 +7,8 @@ __version__ = "0.1.0"
 from openassay.acceptance import AcceptanceResult, LevelStats, run_acceptance
 from openassay.backcalc import BackCalcResult, Sample, back_calculate, back_calculate_many
 from openassay.curve import CalibrationResult, StandardCurve, fit_standard_curve
+from openassay.ingest import read_plate
+from openassay.plate import PlateData, PlateLayout, PlateWell, Well
 from openassay.range import RangeResult, determine_lloq_uloq
 from openassay.report import report_run
 
@@ -16,12 +18,17 @@ __all__ = [
     "CalibrationResult",
     "LevelStats",
     "RangeResult",
+    "PlateData",
+    "PlateLayout",
+    "PlateWell",
     "Sample",
     "StandardCurve",
+    "Well",
     "back_calculate",
     "back_calculate_many",
     "fit_standard_curve",
     "determine_lloq_uloq",
+    "read_plate",
     "report_run",
     "run_acceptance",
 ]